creates menu group stories
